_query: Re-raise the errors.
Though we were adding additionally logging to the errors we were
catching we weren't re-riasing them. Since we're a library we shouldn't
just handle and silence the errors that way as someone using us
then wouldn't be able to act on those exceptions.

Perhaps we need to wrap those into our own exceptions instead of
re-raising the ones from requests.