pypuppetdb/utils: Remove pytz.
Removing the dependency on pytz. It's not needed and adds a rather heavy
library for us to install. Instead we now use a UTC tzinfo class as
shown in the Python docs.

Since PuppetDB always returns UTC timestamps and pypuppetdb itself
wasn't doing anything else but creating a timezone-aware datetime this
achieves the same thing.

It is now up to the user to additionally intstall pytz if the user needs
to work with the datetime objects beyond UTC.

Fixes #9