api: Stop being overprotective of _url/_query.
We had a whole if-statement to figure out stupid uses of path in cases
where it shouldn't occur, such as the 'fact-names' endpoint or the
experimental endpoints.

However, if you do this as a user, that's on you, should've looked at
the API documentation. Both url and query are _methods for a reason.