pypuppetdb: Preliminary v3 support.
Based on the work of @jtopjian this commit adds basic v3 support. It does
not yet expose or use all the new features of v3 API, it just ensures
that what we could do with v2 can now also be done over v3.

Due to changes in PuppetDB the /experimental endpoints no longer exist,
this commit does the necessary surgery to remove that functionality.
This means that when using the v2 API we can no longer access Reports
and Events.

Relates to: #12