puppetdb-3.0/types: Removing the underscores from the Report object's metrics and logs attributes
These additional characters were not necessary

Removing the events attributes primarily because the timestamp attribute
from each event is in a different format than the rest of the return values.
I cannot figure out how to get python to parse a timestamp string
with a colon in the timezone, no documentation suggests that datetime
can support parsing timezones with colons.

The immediate workaround is to query the events endpoint and give
the query parameter report with the current report's hash value.