puppetdb-3.0/types: Enriching the Report object with an events() function
This additional function queries the events endpoint to find all
events associated with the current report and returns a generator
of Event objects.

This required passing the api object as a constructor argument for
this object, modified v3.py and v4.py to pass themselves to the object
to give it access to query this information.

v4 of the Query API already includes a 'resource_events' object in
its response format, I would prefer to preload those events into the
Report object in the constructor but it is in a different format
than what is expected. In addition to the fact that it may run the
risk of Out-of-Memory problems on large reports