pypuppetdb/utils: Adding a new utility function versioncmp
As the name suggests it provides a standard way for pypuppetdb code
ro run standard version comparisons.

The initial intent of this function was to check the Version of PuppetDB
running and check if it is version 3.0 or higher and throw an error
if it failed. In attempting to do so I broke all existing automated
test cases, will have to implement that check in a different way.