pypuppetdb/api/__init__/facts: Replacing the path extensions with their query string equivalents
This resolves https://github.com/voxpupuli/pypuppetdb/issues/67

Instead of appending the fact name and/or the fact value to the path of the facts
call replace it with their query string equivalents. This is prefered over importing
urllib.quote() to escape the strings.

This was reported when a user tried to use the old way to query a MAC
address value and received a 404 from PuppetDB.