pypuppetdb/api/__init__.py: User suggested optimization with optional Report and Catalog attributes
Replacing the Try Except block around the Report code_id, catalog_uuid and
cached_catalog_status attributes with dictionary get(). As per documentation,
https://docs.python.org/2/library/stdtypes.html#dict.get:
Return the value for key if key is in the dictionary, else default. If default is not given, it defaults to None, so that this method never raises a KeyError.

Adding the same change for the Catalog code_id and catalog_uuid attributes.