pypuppetdb/types.py: Deprecating the events() function of the Report object with an events variable.
This events variable is a list of Event objects that are by default
available in a Report returned from PuppetDB. Loading this information
at the beginning as opposed to querying PuppetDB for it when needed.

This sacrifices memory in favour of being infinitely iterable. The
primary application for this is to refactor the PuppetBoard Status Counts
processing.