pypuppetdb/types.py: Removing :class:`pypuppetdb.types.Report` events
This fixes https://github.com/voxpupuli/pypuppetdb/issues/81
This undoes https://github.com/voxpupuli/pypuppetdb/pull/76

This new functionality was not thoroughly tested, the main symtom is that
the timestamps returned by the expanded resource events data are in a
format that Python cannot parse. Namely tiemzones are returned with -04:00
and Python 2.x cannot parse that, not sure about Python 3.