pypuppetdb/QueryBuilder.py: Adding ExtractOperator object.
Following the documentation available https://docs.puppet.com/puppetdb/4.1/api/query/v4/ast.html#extract

This object allows users to create, as the name suggests, an extraction query
to return the number of fields returned in the PuppetDB response. This is
done by passing a string or an array of strings to the add_field(), should
add this to the constructor too.

Adding an option query can be done through the add_query() fuction.
A group_by, https://docs.puppet.com/puppetdb/4.1/api/query/v4/ast.html#groupby,
can be added through add_group_by(). The values in this variable will always
be appended to the end of the ExtractOperator output.

Adding test cases to cover the new code.