pypuppetdb/api.py: Fixing last_report calculation bug causing all nodes to report unreported
This fixes #95

One of the intended code cleanups from this release was to combine the
report_timestamp datetime variable into a single line. This bug unintentionally
caused all nodes in the result set to report a status of unreported.

Also caching the `datetime.now()` code to outside the loop, calling it only
once instead of once-per-node.