pypuppetdb/api.py: POST query in request body
A very large query in the querystring of POST requests can
result in HTTP 414 errors returned from PuppetDB. Putting
the same query into the body of the request will
successfully send the query to the PuppetDB.

An additional flag has been added to specify the POST should
be sent in the body, in order to not break any code that
currently relies on the existing behavior of sending POST in
the querystring.

The body must be a JSON encoded string, and QueryBuilder
objects are not JSON encodable, so if sending the POST in
the body is specified, we need to make sure the query is a
string before trying to encode the payload.

Co-authored-by: Bill Ward <bill@admintome.com>