teststests/test_baseapi.py: POST test support py3
Fix python3 tests for POSTing to PuppetDB with the query
in the body. In python2, requests encodes the contents of
the request body as a string, but in python3, the body is
encoded as a bytes object.

Use the six.b() function to use the correct type when
comparing the expected value with the value that was sent.

Co-authored-by: Bill Ward <bill@admintome.com>