QueryBuilder.py: Use native data structures for internal representation
The [PuppetDB documentation][1] says that queries are JSON arrays, so
this commit updates all of the QueryBuilder classes to use lists/dicts
for the internal representation and defers the JSON serialization until
the `str()` call. This avoids having to manually quote things or do
string manipulation. `obj.json_data()` will return the native data
structure and `str(obj)` will return the serialized JSON.

Most of the changes to the tests are just fixing whitespace, but there
are a few places where duplicated quotes and backslash escaping were
fixed.

[1]: https://puppet.com/docs/puppetdb/5.1/api/query/v4/facts.html#url-parameters

Fixes: #140

Signed-off-by: Andrew Gunnerson <andrew.gunnerson@us.ibm.com>