Switch to Metrics API v2
The v2 (Jolokia) metrics have been available for a while. They are
in Puppet Server 5.3 and 6.9 so anything fairly recent is covered.
The endpoint seems to be identical to v1 metrics, if you just strip
away some additional information that gets returned.

This change is most important for Puppet Server 5.3.12 and 6.9.1
compatibility, released on 10 March 2020. Those releases fully
disable the deprecated v1 metrics endpoints due to CVE-2020-7943.

Closes voxpupuli/pypuppetdb#172.