Enable BaseAPI to be used with 'with' statement
By implementing those two functions, the BaseAPI object can be used with
a 'with' statement like this:

    with pypuppetdb.connect() as session:
        # ...

This will ensure that all HTTP connections are explicitly closed when
exiting the "with" context, thus avoiding possible hangs because of the
open connects.

As stated in the comment, this can happen if our HTTP connections are
sent through a tunnel (for example an ssh tunnel) which will then wait
for connections to close before closing the tunnel.

Explicitly closing HTTP connections will indicate to any tunnel that
we're done with our work and that the tunnel can also be closed in turn.