create disconnect() method for BaseAPI
This new method can help ppl tearing things down without using the
"with" statement. Similarly to the motivation behind implmenting the
"with" statement, this can be useful for connections that were made
through a tunnel (for example ssh).

This method makes it possible for example to have a BaseAPI object
wrapped up in another class, and tear down connections only when that
wrapping class is all done with calls to the object.