Make rtd docs versioned
thanks to a change in the release process:
use tags without the "v" prefix for releases.

To make it easier to remember document the
process and introduce a tool for it - tbump.

Note that we keep the "v" prefix in the
GitHub releases as there is no need
to change this.